Plugin // improve doCleanUpDirectories by removing the language root dir. TranslatablePackage // removed TranslatablePackageTrait and moved code into class.
